+++ src/main/java/org/petmarket/users/service/UserAuthService.java
@@ -3,7 +3,7 @@ package org.petmarket.users.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.petmarket.errorhandling.ItemNotFoundException;
-import org.petmarket.users.converter.UserDtoMapper;
+import org.petmarket.users.mapper.UserDtoMapper;
 import org.petmarket.users.dto.UserRequestDto;
 import org.petmarket.users.dto.UserResponseDto;
 import org.petmarket.users.entity.Role;
